backend/requirements.txt
asgiref==3.11.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
databricks==0.2
databricks-sdk==0.73.0
databricks-sql-connector==4.2.1
Django==5.2.8
et_xmlfile==2.0.0
google-auth==2.43.0
idna==3.11
lz4==4.4.5
numpy==2.3.5
oauthlib==3.3.1
openpyxl==3.1.5
pandas==2.2.3
protobuf==6.33.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
pyspark==3.5.6
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
sqlparse==0.5.4
thrift==0.20.0
tzdata==2025.2
urllib3==2.5.0
python-dotenv==1.2.1
